requirements.txt
Django>=5.0
djangorestframework>=3.14
whitenoise>=6.6
psycopg[binary]>=3.2
dj-database-url>=2.2
gunicorn>=21.2